Non-linear parameter estimation of Helmert
Renamed the existing Helmert7Param class to `Helmert7ParamLinear` and
added `Helmert7ParamNonLinear`. The latter can estimate parameters based
on the full rotation matrix. This requires a different approach which
can be summed up as

1. Center coordinates around their centroids (x0, y0) before solving - this avoids
   numerical issues with the large geodetic coordinates (~3-5 million)

2. Use scipy.optimize.least_squares instead of the manual Gauss-Newton iteration - it's
   more robust for this non-linear problem

3. Convert translation parameters back to original coordinate system after solving:
   t_orig = t_norm - k * R @ x0 + y0

Step two will find a solution iteratively and is likely to be more
precise in more cases.

Documentation and tests added.